requirements.txt
aiogram==3.3.0
requests
beautifulsoup4
lxml
sqlalchemy
python-dotenv
schedule
